moved column name functions to parsing_tools.py
