+++ src/EC_MS/Combining.py
@@ -15,6 +15,7 @@ import re
 
 # import os #, sys
 import numpy as np
+from .parsing_tools import get_timecol, is_time, get_type, timestamp_to_seconds
 
 # import warnings # actually not that useful, as you loose information about
 # when in the running of the code the problem appeared.
@@ -1087,147 +1088,6 @@ def parse_CHI_header(data):
         data["Ns"][mask] = i
         ti = ti + dt
         i = i + 1
-
-
-def is_time(col, data=None, verbose=False):
-    """
-    determines if a column header is a time variable, 1 for yes 0 for no
-    """
-    if verbose:
-        print("\nfunction 'is_time' checking '" + col + "'!")
-    timecol = get_timecol(col, data, verbose=verbose)
-    return timecol == col[: len(timecol)]  # so that a * at the end doesn't mess it up
-
-
-def is_MS_data(col):
-    if re.search(r"^M[0-9]+-[xy]", col):
-        return True
-    return False
-
-
-def is_EC_data(col):
-    from .EC import EC_cols_0
-
-    if col in EC_cols_0:
-        # this list should be extended as needed
-        return True
-    if col is None:
-        return False
-    if col[-1] == "*" and col[:-1] in EC_cols_0:
-        return True
-    return False
-
-
-def is_Xray_data(col):
-    if is_EC_data(col):
-        return False
-    if is_MS_data(col):
-        return False
-    return True
-
-
-def get_type(col, dataset=None):
-    if dataset is not None:
-        if "col_types" in dataset and col in dataset["col_types"]:
-            return dataset["col_types"][col]
-        elif "data_type" in dataset and dataset["data_type"] in [
-            "EC",
-            "MS",
-            "SI",
-            "SPEC",
-        ]:
-            return dataset["data_type"]
-    if is_EC_data(col):
-        return "EC"
-    if is_MS_data(col):
-        return "MS"
-    elif col[-2:] in ["-x", "-y"]:
-        return "cinfdata"  # it's cinfdata but not from a mass channel
-    if col is None:
-        return None
-    print(
-        "WARNING: "
-        + col
-        + " is not recognized. Assuming type 'EC'.\n "
-        + " Consider adding to EC_cols_0 in EC.py if so."
-    )
-    return "EC"  #'Xray' # to be refined later...
-
-
-def get_cols_for_mass(mass, dataset=None):
-    """
-    Eventually this might make the 'rename_<format>_cols' functions obsolete.
-    """
-    if dataset is None:
-        xcol, ycol = mass + "-x", mass + "-y"
-    elif "mass_cols" in dataset:
-        xcol, ycol = dataset["mass_cols"][mass]
-    else:
-        ycol = mass + "-y"
-        xcol = get_timecol(ycol, dataset)
-    return xcol, ycol
-
-
-def get_timecol(col=None, dataset=None, data_type=None, verbose=False):
-    # print('getting timecol for ' + col + '. datset = ' + (str(dataset)+ 20*' ')[:20]) # debugging
-    if dataset is not None and "timecols" in dataset and col in dataset["timecols"]:
-        return dataset["timecols"][col]
-
-    if data_type is None:
-        data_type = get_type(col, dataset)
-
-    if data_type == "EC":
-        return "time/s"
-    elif data_type == "MS":
-        if col is None:
-            return (
-                "M32-x"  # probably the least likely timecol to be missing from MS data
-            )
-        else:
-            return col[:-2] + "-x"
-    elif data_type == "SI":
-        return col.split(" - ")[0] + " - Time [s]"
-    elif data_type == "RGA":
-        return "Time(s)"
-    elif data_type == "Xray":
-        return "t"  # to be refined later...
-    elif data_type == "CHI":
-        if dataset is not None and "Time/sec" in dataset:
-            return "Time/sec"
-        else:
-            return "Time/s"
-    elif col[-2:] in ["-y", "-x"]:  # a timecol is its own timecol
-        return col[:-2] + "-x"  # for any data downloaded from cinfdata
-    else:
-        print(
-            "couldn't get a timecol for " + str(col) + ". data_type=" + str(data_type)
-        )
-        return None
-
-
-def timestamp_to_seconds(timestamp):
-    """
-    seconds since midnight derived from timestamp hh:mm:ss
-    """
-    h = int(timestamp[0:2])
-    m = int(timestamp[3:5])
-    s = int(timestamp[6:8])
-    seconds = 60 ** 2 * h + 60 * m + s
-    return seconds
-
-
-def seconds_to_timestamp(seconds):
-    """
-    timestamp hh:mm:ss derived from seconds since midnight
-    """
-    h = int(seconds / 60 ** 2)
-    seconds = seconds - 60 ** 2 * h
-    m = int(seconds / 60)
-    seconds = seconds - 60 * m
-    s = int(seconds)
-    timestamp = "{0:2d}:{1:2d}:{2:2d}".format(h, m, s)
-    timestamp = timestamp.replace(" ", "0")
-    return timestamp
 
 
 def dayshift(dataset, days=1):
